Remove Decoder
Closes #27
